+++ src/components/Tips.tsx
